Add some aliases for archiving
Occasionally we want to remember a commit for a while but aren't intending
to merge it in or use it for the foreseeable future. We don't want to lose
it, but stashing it will bloat the stash, and committing it to a local
branch will bloat the output of 'git branch -l'. Pushing it to a remote
and deleting the local branch makes local branch output nicer but bloats
the remote - and we might not necessarily have a remote anyway.

These aliases use some plumbing commands to create new refs pointing at a
specific location - under refs/archive, so that they don't show up as a
local or remote branch, but so that they still hold a reference to the
commit and prevent it from becoming unreachable.