Add alias for intent-to-add
Git doesn't allow interactive/patch adding of untracked files, since all
it would be able to do is present one giant hunk of the whole file and
you'll need to edit it vim anyway.

However, occasionally this is what we want to do, so add a quick alias for
the --intent-to-add option to git add, so that we can `git an` and then
`git ap` the new file to start editing the hunks.