Add some pull-request aliases
Occasionally it's useful to grab a pull request from github and check it
out locally, for example to check that the tests pass. Similarly, we may
want to grab github's autogenerated merge result and test that. Obviously
we could add the user's fork as a new remote and pull their branch, but
that makes more admin on the remotes side of things.

Add some aliases for grabbing pull requests and their merge results and
creating local refs to them, either as local branches or in refs/pr.