Add alias for applying archived commits
Most of the existing aliases for archived commits are predicated on the
use case of a line of work that cannot be currently merged but which might
be resurrected in the future, and in the interim we don't want it
polluting our branch list.

There is another common use case, though: a set of local changes (for
example, to configuration files) that we often want to apply locally but
don't want to commit to the actual repo. This use case is often a place
when we might use git stash - stash the local config changes, switch
branch, apply the stashed changes.

This new alias is for this case - if we have a commit in our archive which
is, say, a local configuration change, we can now easily apply it on top
of our working directory.